[REFACTOR]: 별점 컴포넌트 공통부분 분리
